블로그 및 타일틀 컴포넌트로 분리
