+++ src/Title.jsx
@@ -1,0 +1,11 @@
+function Title() {
+  return (
+    <>
+    <div className='black-bg'>
+      React + Vite로 만드는 블로그
+    </div>
+    </>
+  );
+}
+
+export default Title;
